move generateErrorSummary
