+++ app/pages/sections/catalogue-solutions/order-item/contextCreator.js
@@ -1,7 +1,7 @@
 import { baseUrl } from '../../../../config';
 import { generateErrorMap } from '../../../../helpers/contextCreators/generateErrorMap';
 import { generateQuestions } from '../../../../helpers/generateQuestions';
-import { generateErrorSummary } from '../../../../helpers/generateErrorSummary';
+import { generateErrorSummary } from '../../../../helpers/contextCreators/generateErrorSummary';
 import { generateAddPriceTable } from '../../../../helpers/contextCreators/generateAddPriceTable';
 
 export const getContext = ({
